refactor(desktop): isolate model boundary types
