+++ apps/desktop/src/features/library-workspace/records/model/filterTypes.ts
@@ -1,0 +1,14 @@
+export type RecordExplorerFilterTag = {
+  id: string;
+  name: string;
+};
+
+export type RecordExplorerFilterGroup = {
+  key: string;
+  label: string;
+  tags: readonly RecordExplorerFilterTag[];
+};
+
+export type RecordExplorerSelectedFilters = Readonly<Record<string, readonly string[]>>;
+
+export type SelectedRecordFilters = Record<string, string[]>;
